Clean: move some funcs from splitter.c to splitter_utils.c
