+++ utils/splitter.c
@@ -3,87 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   splitter.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aldokezer <aldokezer@student.42.fr>        +#+  +:+       +#+        */
+/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 10:24:52 by mbartos           #+#    #+#             */
-/*   Updated: 2024/03/01 22:35:42 by aldokezer        ###   ########.fr       */
+/*   Updated: 2024/03/02 13:17:13 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// #include "../minishell.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-
-int	num_of_strings(char *str, char c)
-{
-	int	i;
-	int	n_words;
-
-	n_words = 0;
-	i = 0;
-	if (str[0] && str[0] != c)
-		n_words++;
-	while (str[i])
-	{
-		if (str[i] == c && (str[i + 1] != c && str[i + 1] != 0))
-			n_words++;
-		i++;
-	}
-	return (n_words);
-}
-
-char	*str_add(char *str, char c, int *index)
-{
-	int		i;
-	int		length;
-	char	*word;
-
-	length = 0;
-	while (str[length] && str[length] != c)
-		length++;
-	word = malloc(sizeof(char) * (length + 1));
-	if (word == NULL)
-		return (NULL);
-	i = 0;
-	*index = *index + length;
-	while (i < length)
-	{
-		word[i] = str[i];
-		i++;
-	}
-	word[i] = 0;
-	return (word);
-}
-
-char	**ft_split(char const *s, char c)
-{
-	char	*str;
-	char	**array;
-	int		i;
-	int		j;
-
-	str = (char *) s;
-	array = (char **) malloc(sizeof(char *) * (num_of_strings(str, c) + 1));
-	if (array == NULL)
-		return (NULL);
-	i = 0;
-	j = 0;
-	if (str[i] && str[i] != c)
-		array[j++] = str_add(&str[i], c, &i);
-	while (str[i])
-	{
-		if (str[i] == c && (str[i + 1] != c && str[i + 1] != 0))
-			array[j++] = str_add(&str[i + 1], c, &i);
-		i++;
-	}
-	array[j] = NULL;
-	return (array);
-}
-
-
-
+#include "../minishell.h"
 
 // SPLITTER
 /*
@@ -105,46 +32,6 @@ Parse by <, >, |, <<, >> (if not in quotes)
 - stejne pokud najdu '
 
 */
-
-
-int	is_whitespace(char c)
-{
-	if (c == 9 || c == 10 || c == 12 || c == 13 || c == 32)
-		return (1);
-	return (0);
-}
-
-int	is_quote(char c)
-{
-	if (c == '\'' || c == '\"')
-		return (1);
-	return (0);
-}
-
-int	is_token(char c)
-{
-	if (c == '|' || c == '<' || c == '>')
-		return (1);
-	return (0);
-}
-
-int	is_end_of_word(char	c)
-{
-	if (is_whitespace(c) || is_quote(c) || is_token(c) || c == '\0')
-		return (1);
-	return (0);
-}
-
-size_t	word_length(char *str)
-{
-	size_t	count;
-
-	count = 0;
-	while(!is_end_of_word(str[count]))
-		count++;
-	return(count);
-}
-
 char	*handle_redirect_input(char *str, size_t *index)
 {
 	size_t	size_to_malloc;
@@ -273,7 +160,7 @@ char	*handle_word_outside_quotes(char *str, size_t *index)
 	return (str_out);
 }
 
-void	split2 (char *line)
+void	splitter(char *line)
 {
 	char	*array_of_words[30];
 	size_t	i;
@@ -338,16 +225,16 @@ void	split2 (char *line)
 // cc splitter.c -Wall -Wextra -Werror -L/opt/homebrew/opt/readline/lib -lreadline
 
 
-// int	main(void)
-// {
-// 	char	*line;
-// 	// char line[] = "a | b";
+int	main(void)
+{
+	char	*line;
+	// char line[] = "a | b";
 
-// 	line = readline("Write command: ");
-// 	printf("What I got:    %s\n", line);
-// 	split2(line);
+	line = readline("Write command: ");
+	printf("What I got:    %s\n", line);
+	splitter(line);
 
-// 	// char	*args[] = {"/bin/echo", "\\n", NULL};
-// 	// execve(args[0], args, NULL);
-// 	return (0);
-// }
+	// char	*args[] = {"/bin/echo", "\\n", NULL};
+	// execve(args[0], args, NULL);
+	return (0);
+}
